[fix] Radios not checked on load
Radios were loaded too fast so when checking if they should be checked by default or not, the value was `undefined`

Using a state `isDataLoaded` to check if all the data was loaded or not, it fixes the problem. Probably not the prettiest fix ever made, but it does the job at the moment.